requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
joblib
country_converter
kafka-python
mysql-connector-python
python-dotenv
dash
dash-bootstrap-components
plotly
statsmodels
